requirements/local.txt
bcrypt==3.1.7
sqlparse==0.3.1
mysqlclient==1.4.6
PyJWT==1.7.1
django-cors-headers==3.3.0


